requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.32.0
pypdf>=5.0.0
openpyxl>=3.1.0
PyYAML>=6.0.0
